refactor(api,mcp,loadtest): switch to /me self-mint, drop platform-admin endpoint
The original /api/v1/admin/load-test/mcp-token endpoint (PR #2136)
gated load-test bearer minting at the platform_admin tier because the
endpoint took a workspaceId in the body and could mint cross-tenant.
That gave platform_admin a new MCP-via-tool-call data exfiltration
path: mint a token for any workspace, then run executeSQL against
that workspace's data via the agent loop. Audit + rate limit made it
traceable, not prevented.

For the actual use case (CI minting a token for the dedicated
load-test workspace it owns), the platform-admin tier is over-secured
— self-mint is functionally equivalent to a workspace member running
the OAuth 2.1 ceremony against their own workspace. Same scope, same
audience, same workspace claim. The only difference is bypassing the
consent ceremony.

This PR closes the cross-tenant surface entirely:

- Add POST /api/v1/me/load-test/mcp-token (workspace-member gated,
  scoped to caller's active workspace, no body workspaceId — by
  construction, you cannot mint for a workspace you don't own).
- Delete POST /api/v1/admin/load-test/mcp-token. Pre-launch, no
  external callers, clean breaking change.
- Switch loadtest.sh BASE_URL default to https://mcp.useatlas.dev
  (the brand surface — k6 then exercises exactly what real customers
  hit). Region binding is implicit in BASE_URL; no body region field.
- Move docs page from platform-ops/ to guides/ (it's no longer an
  operator-only surface).
- Audit catalog comment updated: scope is now `workspace`, not
  `platform`, so audit queries filtering by org_id pick these up.
- Tests: drop admin-load-test.test.ts (15 cases), add
  me-load-test.test.ts (15 cases) covering the new shape +
  workspaceId-injection guard (body field is ignored, response always
  carries the session's active org).

OpenAPI + api-reference MDX regenerated.

Closes #2135 (the follow-up that should have been the original
design).
