+++ apps/docs/content/docs/api-reference/me-load-test/meta.json
@@ -1,0 +1,3 @@
+{
+  "title": "Me — Load Test"
+}
