+++ apps/docs/content/docs/api-reference/me-load-test/postMeLoadTestMcpToken.mdx
@@ -1,0 +1,28 @@
+---
+title: Mint a short-lived MCP-scoped JWT for the caller's own workspace
+full: true
+_openapi:
+  method: POST
+  toc: []
+  structuredData:
+    headings: []
+    contents:
+      - content: >-
+          Mints a JWT signed by the regional Better Auth JWKS, scoped to
+          `mcp:read` and bound to the caller's active workspace. Functionally
+          equivalent to a workspace member completing the OAuth 2.1 ceremony
+          against their own workspace, but skipping the consent flow — useful
+          for load tests, smoke tests, and any flow that needs an MCP bearer
+          programmatically. TTL is capped at 3600s server-side; requests above
+          the cap are rejected with 400. The bearer must NEVER be logged or
+          persisted; the audit row carries only `jti` for forensic correlation.
+          The token's audience is bound to this regional API. Calling this
+          against `mcp.useatlas.dev` and using the bearer against
+          `mcp-eu.useatlas.dev` will fail with 401 (audience mismatch).
+---
+
+{/* This file was generated by Fumadocs. Do not edit this file directly. Any changes should be made by running the generation command again. */}
+
+Mints a JWT signed by the regional Better Auth JWKS, scoped to `mcp:read` and bound to the caller's active workspace. Functionally equivalent to a workspace member completing the OAuth 2.1 ceremony against their own workspace, but skipping the consent flow — useful for load tests, smoke tests, and any flow that needs an MCP bearer programmatically. TTL is capped at 3600s server-side; requests above the cap are rejected with 400. The bearer must NEVER be logged or persisted; the audit row carries only `jti` for forensic correlation. The token's audience is bound to this regional API. Calling this against `mcp.useatlas.dev` and using the bearer against `mcp-eu.useatlas.dev` will fail with 401 (audience mismatch).
+
+<APIPage document={"./openapi.json"} operations={[{"path":"/api/v1/me/load-test/mcp-token","method":"post"}]} />
