+++ apps/docs/content/docs/api-reference/meta.json
@@ -57,6 +57,7 @@
     "demo",
     "discord",
     "health",
+    "me-load-test",
     "me-oauth-clients",
     "mode",
     "onboarding",
@@ -64,7 +65,6 @@
     "platform-admin",
     "platform-admin-backups",
     "platform-admin-custom-domains",
-    "platform-admin-load-test",
     "platform-admin-residency",
     "platform-admin-sla",
     "platform-plugin-catalog",
