--- apps/docs/content/docs/api-reference/platform-admin-load-test/meta.json
@@ -1,3 +1,0 @@
-{
-  "title": "Platform Admin — Load Test"
-}
