--- apps/docs/content/docs/api-reference/platform-admin-load-test/postAdminLoadTestMcpToken.mdx
@@ -1,28 +1,0 @@
----
-title: Mint a short-lived MCP-scoped JWT for load testing
-full: true
-_openapi:
-  method: POST
-  toc: []
-  structuredData:
-    headings: []
-    contents:
-      - content: >-
-          Mints a JWT signed by the regional Better Auth JWKS, scoped to
-          `mcp:read` and bound to the requested workspace. Intended for the CI
-          load-test workflow (#2129) — k6 calls this once per run, then carries
-          the bearer in `Authorization: Bearer <token>` against
-          `/mcp/{workspaceId}/sse`. TTL is capped at 3600s server-side; requests
-          above the cap are rejected with 400. The bearer must NEVER be logged
-          or persisted; the audit row carries only `jti` for forensic
-          correlation. Region constraint: the token's audience is bound to one
-          regional API. Calling this against `api.useatlas.dev` and using the
-          bearer against `api-eu.useatlas.dev` will fail with 401 (audience
-          mismatch).
----
-
-{/* This file was generated by Fumadocs. Do not edit this file directly. Any changes should be made by running the generation command again. */}
-
-Mints a JWT signed by the regional Better Auth JWKS, scoped to `mcp:read` and bound to the requested workspace. Intended for the CI load-test workflow (#2129) — k6 calls this once per run, then carries the bearer in `Authorization: Bearer <token>` against `/mcp/{workspaceId}/sse`. TTL is capped at 3600s server-side; requests above the cap are rejected with 400. The bearer must NEVER be logged or persisted; the audit row carries only `jti` for forensic correlation. Region constraint: the token's audience is bound to one regional API. Calling this against `api.useatlas.dev` and using the bearer against `api-eu.useatlas.dev` will fail with 401 (audience mismatch).
-
-<APIPage document={"./openapi.json"} operations={[{"path":"/api/v1/admin/load-test/mcp-token","method":"post"}]} />
