+++ apps/docs/content/docs/guides/meta.json
@@ -6,6 +6,7 @@
     "choosing-an-integration",
     "mcp",
     "mcp-hosted",
+    "mcp-load-test-tokens",
     "slack",
     "python",
     "notebook",
