+++ apps/docs/content/docs/platform-ops/meta.json
@@ -12,7 +12,6 @@
     "abuse-prevention",
     "plugin-catalog",
     "encryption-key-rotation",
-    "plugin-config-residue-audit",
-    "mcp-load-test-tokens"
+    "plugin-config-residue-audit"
   ]
 }
