+++ apps/docs/openapi.json
@@ -17,7 +17,7 @@
   ],
   "components": {
     "schemas": {
-      "MintMcpTokenResponse": {
+      "MintMyMcpTokenResponse": {
         "type": "object",
         "properties": {
           "bearer": {
@@ -25,7 +25,8 @@
             "description": "The signed JWT. Carry verbatim in `Authorization: Bearer <bearer>`. Must NEVER be embedded in logs or stored to disk."
           },
           "workspaceId": {
-            "type": "string"
+            "type": "string",
+            "description": "Resolved from the caller's session active organization."
           },
           "audience": {
             "type": "string",
@@ -41,7 +42,7 @@
           },
           "sub": {
             "type": "string",
-            "description": "Synthetic subject — `loadtest:<workspaceId>:<random>`. Safe to log; appears in every audit row downstream."
+            "description": "Synthetic subject — `loadtest:<workspaceId>:<random>`. Safe to log."
           },
           "scope": {
             "type": "string",
@@ -58,30 +59,17 @@
           "scope"
         ]
       },
-      "MintMcpTokenRequest": {
+      "MintMyMcpTokenRequest": {
         "type": "object",
         "properties": {
-          "workspaceId": {
-            "type": "string",
-            "minLength": 1,
-            "description": "The Atlas organization id the minted token will be scoped to. The MCP path must match (`/mcp/{workspaceId}/sse`)."
-          },
           "ttlSeconds": {
             "type": "integer",
             "minimum": 60,
             "maximum": 3600,
             "default": 300,
-            "description": "Token TTL in seconds. Server-side ceiling is 3600s; default 300s. Requests for larger values are rejected with 400 — the cap is not silently clamped."
-          },
-          "region": {
-            "type": "string",
-            "minLength": 1,
-            "description": "Region key (e.g. `us`, `eu`, `apac`) matching `residency.regions[*]` in atlas.config.ts. Defaults to this API instance's region when omitted. The token's audience is bound to the resolved region's API URL — verifying against a different regional MCP edge will fail closed."
+            "description": "Token TTL in seconds. Server-side ceiling is 3600s; default 300s. Requests for larger values are rejected with 400."
           }
-        },
-        "required": [
-          "workspaceId"
-        ]
+        }
       }
     },
     "parameters": {},
@@ -56291,19 +56279,19 @@
         "operationId": "deleteAdminSettingsByKey"
       }
     },
-    "/api/v1/admin/load-test/mcp-token": {
+    "/api/v1/me/load-test/mcp-token": {
       "post": {
         "tags": [
-          "Platform Admin — Load Test"
+          "Me — Load Test"
         ],
-        "summary": "Mint a short-lived MCP-scoped JWT for load testing",
-        "description": "Mints a JWT signed by the regional Better Auth JWKS, scoped to `mcp:read` and bound to the requested workspace. Intended for the CI load-test workflow (#2129) — k6 calls this once per run, then carries the bearer in `Authorization: Bearer <token>` against `/mcp/{workspaceId}/sse`. TTL is capped at 3600s server-side; requests above the cap are rejected with 400. The bearer must NEVER be logged or persisted; the audit row carries only `jti` for forensic correlation. Region constraint: the token's audience is bound to one regional API. Calling this against `api.useatlas.dev` and using the bearer against `api-eu.useatlas.dev` will fail with 401 (audience mismatch).",
+        "summary": "Mint a short-lived MCP-scoped JWT for the caller's own workspace",
+        "description": "Mints a JWT signed by the regional Better Auth JWKS, scoped to `mcp:read` and bound to the caller's active workspace. Functionally equivalent to a workspace member completing the OAuth 2.1 ceremony against their own workspace, but skipping the consent flow — useful for load tests, smoke tests, and any flow that needs an MCP bearer programmatically. TTL is capped at 3600s server-side; requests above the cap are rejected with 400. The bearer must NEVER be logged or persisted; the audit row carries only `jti` for forensic correlation. The token's audience is bound to this regional API. Calling this against `mcp.useatlas.dev` and using the bearer against `mcp-eu.useatlas.dev` will fail with 401 (audience mismatch).",
         "requestBody": {
-          "required": true,
+          "required": false,
           "content": {
             "application/json": {
               "schema": {
-                "$ref": "#/components/schemas/MintMcpTokenRequest"
+                "$ref": "#/components/schemas/MintMyMcpTokenRequest"
               }
             }
           }
@@ -56314,13 +56302,13 @@
             "content": {
               "application/json": {
                 "schema": {
-                  "$ref": "#/components/schemas/MintMcpTokenResponse"
+                  "$ref": "#/components/schemas/MintMyMcpTokenResponse"
                 }
               }
             }
           },
           "400": {
-            "description": "Invalid body, ttl out of range, or unknown region",
+            "description": "Invalid body, ttl out of range, or no active workspace on the session",
             "content": {
               "application/json": {
                 "schema": {
@@ -56346,17 +56334,6 @@
           },
           "401": {
             "description": "Authentication required",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "object",
-                  "additionalProperties": {}
-                }
-              }
-            }
-          },
-          "403": {
-            "description": "Caller is not platform_admin (or MFA not enrolled)",
             "content": {
               "application/json": {
                 "schema": {
@@ -56392,7 +56369,7 @@
             }
           },
           "429": {
-            "description": "Per-endpoint mint rate exceeded (10/min/admin)",
+            "description": "Per-endpoint mint rate exceeded (10/min/caller)",
             "content": {
               "application/json": {
                 "schema": {
@@ -56417,7 +56394,7 @@
             }
           },
           "500": {
-            "description": "Internal error during minting (e.g. JWK decryption failed)",
+            "description": "Internal error during minting",
             "content": {
               "application/json": {
                 "schema": {
@@ -56467,7 +56444,7 @@
             }
           }
         },
-        "operationId": "postAdminLoadTestMcpToken"
+        "operationId": "postMeLoadTestMcpToken"
       }
     },
     "/api/v1/platform/workspaces": {
@@ -65193,6 +65170,9 @@
       "name": "Health"
     },
     {
+      "name": "Me — Load Test"
+    },
+    {
       "name": "Me — OAuth Clients"
     },
     {
@@ -65212,9 +65192,6 @@
     },
     {
       "name": "Platform Admin — Custom Domains"
-    },
-    {
-      "name": "Platform Admin — Load Test"
     },
     {
       "name": "Platform Admin — Residency"
