+++ packages/api/src/api/index.ts
@@ -309,18 +309,17 @@ try {
   );
 }
 
-// Platform-admin MCP load-test token mint (#2135). Lives at
-// /api/v1/admin/load-test even though the gate is platform_admin: the
-// surface is operator-facing tooling, not workspace-admin functionality,
-// and the path mirrors the issue spec so CI workflows are stable.
+// Self-mint MCP load-test token (#2135 follow-up). Workspace-member
+// gated, scoped to the caller's active workspace. Replaces the original
+// platform-admin variant — see me-load-test.ts header for the rationale.
 try {
-  const { adminLoadTest } = await import("./routes/admin-load-test");
-  app.route("/api/v1/admin/load-test", adminLoadTest);
-  log.info("Admin load-test routes enabled");
+  const { meLoadTest } = await import("./routes/me-load-test");
+  app.route("/api/v1/me/load-test", meLoadTest);
+  log.info("Me load-test routes enabled");
 } catch (err) {
   log.error(
     { err: err instanceof Error ? err : new Error(String(err)) },
-    "Failed to load admin load-test routes",
+    "Failed to load me load-test routes",
   );
 }
